+++ CHANGELOG.md
@@ -6,6 +6,12 @@
   a new tab
 * **Admin Feature** Ability to add ESTA to entity Added
 * Trail map code updated
+* Added API for customer graphs
+* Revamped graphs code
+* Added API for customer local purchase pie charts
+* Added API for customer snippets
+* Added API for sector purchase list for customer dashboard
+* **Admin Feature** Fixed org sector on user edit layout and text
 
 # v0.9.7